remove unused classes refactor packages
