+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
         <!-- android:clearTaskOnLaunch="true" is a workaround to prevent launching without registered player -->
         <activity
             android:clearTaskOnLaunch="true"
-            android:name=".LoginActivity"
+            android:name=".layout.LoginActivity"
             android:label="@string/app_name"
             android:windowSoftInputMode="adjustResize">
             <intent-filter>
@@ -32,23 +32,23 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".QuestionActivity"
+            android:name=".layout.QuestionActivity"
             android:label="@string/app_name"
             android:configChanges="orientation|screenSize"
             />
         <activity
-            android:name=".CategorySelectionActivity"
+            android:name=".layout.CategorySelectionActivity"
             android:label="@string/app_name"
             android:configChanges="orientation|screenSize"
              />
         <activity
-            android:name=".BottomNavigation"
+            android:name=".layout.BottomNavigation"
             android:label="@string/app_name">
         </activity>
-        <activity android:name=".MultiplayerLobbyActivity" />
-        <activity android:name=".MultiplayerQuestionActivity" />
+        <activity android:name=".layout.MultiplayerLobbyActivity" />
+        <activity android:name=".layout.MultiplayerQuestionActivity" />
         <activity
-            android:name=".SettingsActivity"
+            android:name=".layout.SettingsActivity"
             android:label="@string/title_activity_settings" />
     </application>
