+++ app/src/main/java/com/app/game/quizee/layout/AppCompatPreferenceActivity.java
@@ -1,4 +1,4 @@
-package com.app.game.quizee;
+package com.app.game.quizee.layout;
 
 import android.content.res.Configuration;
 import android.os.Bundle;
