+++ app/src/main/java/com/app/game/quizee/layout/QuestionActivity.java
@@ -1,4 +1,4 @@
-package com.app.game.quizee;
+package com.app.game.quizee.layout;
 
 import android.annotation.TargetApi;
 import android.content.DialogInterface;
@@ -26,6 +26,7 @@ import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.app.game.quizee.R;
 import com.app.game.quizee.backend.Achievement;
 import com.app.game.quizee.backend.Answer;
 import com.app.game.quizee.backend.Category;
@@ -35,6 +36,7 @@ import com.app.game.quizee.backend.Player;
 import com.app.game.quizee.backend.PlayerManager;
 import com.app.game.quizee.backend.PowerUp;
 import com.app.game.quizee.backend.Question;
+import com.app.game.quizee.backend.TriviaApi;
 
 import java.util.ArrayList;
 import java.util.Collections;
