+++ app/src/main/java/com/app/game/quizee/layout/SettingsActivity.java
@@ -1,4 +1,4 @@
-package com.app.game.quizee;
+package com.app.game.quizee.layout;
 
 
 import android.annotation.TargetApi;
@@ -14,6 +14,7 @@ import android.util.Log;
 import android.view.MenuItem;
 import android.widget.Toast;
 
+import com.app.game.quizee.R;
 import com.app.game.quizee.backend.MusicService;
 import com.app.game.quizee.backend.PlayerManager;
