+++ app/src/main/java/com/app/game/quizee/layout/ShopFragment.java
@@ -1,4 +1,4 @@
-package com.app.game.quizee;
+package com.app.game.quizee.layout;
 
 import android.app.Activity;
 import android.content.Context;
@@ -17,6 +17,7 @@ import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.app.game.quizee.R;
 import com.app.game.quizee.backend.Achievement;
 import com.app.game.quizee.backend.PowerUp;
 import com.app.game.quizee.backend.Player;
