+++ app/src/main/res/layout/activity_bottom_navigation.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.app.game.quizee.BottomNavigation"
+    tools:context="com.app.game.quizee.layout.BottomNavigation"
     android:background="@color/primaryBackground">
 
     <android.support.v7.widget.Toolbar
