+++ app/src/main/res/layout/content_multiplayer_question.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.app.game.quizee.MultiplayerQuestionActivity"
+    tools:context="com.app.game.quizee.layout.MultiplayerQuestionActivity"
     tools:showIn="@layout/activity_multiplayer_question">
