+++ app/src/main/res/layout/fragment_shop.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.app.game.quizee.ShopFragment">
+    tools:context="com.app.game.quizee.layout.ShopFragment">
 
     <TextView
         android:id="@+id/shop_title"
