+++ app/src/main/res/layout/fragment_top_players.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
-    tools:context="com.app.game.quizee.TopPlayersFragment">
+    tools:context="com.app.game.quizee.layout.TopPlayersFragment">
 
 
         <LinearLayout
